refactor, indexContacts
 changed structure, added single action controllers
