+++ repositories/contacts/removeContact.js
